requirements.txt
fastapi
uvicorn
spacy
PyPDF2
networkx
matplotlib
streamlit
requests
pillow
python-multipart
